連続授業の対応

+++ src-tauri/src/algorithm/aco/violations.rs
@@ -1,12 +1,12 @@
 use serde::{Deserialize, Serialize};
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Violations {
-    period: u64,
+    period: usize,
     rooms: Vec<usize>,
 }
 
 impl Violations {
-    pub fn new(period: u64, rooms: Vec<usize>) -> Violations {
+    pub fn new(period: usize, rooms: Vec<usize>) -> Violations {
         Violations { period, rooms }
     }
 }
