+++ src-tauri/src/input/student_group.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct StudentGroup {
-    pub id: u64,
+    pub id: usize,
     pub name: String,
-    pub index: u64,
+    pub index: usize,
 }
